+++ bin/split_kraken_report.py
@@ -17,27 +17,32 @@ def parse_depth(name):
     return depth
 
 
-def save_file(name, lines):
+def save_file(name, lines, header):
     if len(lines) == 0:
         return
     outfile = ".".join([name, "kreport_split.txt"])
     with open(outfile, "w") as f:
+        if header:
+            f.write(header)
         for line in lines:
             f.write(line)
 
 
-def parse_report_file(report_file, split_strings, split_rank, save_json):
+def parse_report_file(report_file, split_strings, split_rank, ignore, save_json):
     depth_dict = {}
     lines = {"remainder": []}
     key = "remainder"
+    ignore_entry = False
     hierarchy = []
     max_depth = -1
     entries = {}
+    header = None
 
     # parses a kraken or bracken file
     with open(report_file, "r") as f:
         for line in f:
             if line.startswith("% of Seqs"):
+                header = line
                 continue
             try:
                 (
@@ -83,6 +88,7 @@ def parse_report_file(report_file, split_strings, split_rank, save_json):
 
             while depth <= max_depth:
                 add_hierarchy = True
+                ignore_entry = False
                 del depth_dict[max_depth]
                 if len(depth_dict) > 0:
                     key = list(depth_dict.values())[-1]
@@ -91,21 +97,29 @@ def parse_report_file(report_file, split_strings, split_rank, save_json):
                     key = "remainder"
                     max_depth = -1
 
-            if name in split_strings or raw_rank == split_rank:
+            if ncbi in ignore:
+                key = name
+                depth_dict[depth] = name
+                max_depth = depth
+                ignore_entry = True
+                add_hierarchy = False
+            elif name in split_strings or raw_rank == split_rank:
                 key = name
                 lines[key] = hierarchy.copy()
                 depth_dict[depth] = name
                 max_depth = depth
+                ignore_entry = False
             elif add_hierarchy:
                 for ancestor in hierarchy:
                     if ancestor not in lines[key]:
                         lines[key].append(ancestor)
 
             hierarchy.append(line)
-            lines[key].append(line)
+            if not ignore_entry:
+                lines[key].append(line)
 
-        for key in lines:
-            save_file(key, lines[key])
+    for key in lines:
+        save_file(key, lines[key], header)
 
     if save_json:
         with open(report_file.replace(".txt", ".json"), "w") as outfile:
@@ -129,6 +143,14 @@ def main():
         default=[],
         help="List of taxon names to split the file by",
     )
+    parser.add_argument(
+            "--ignore",
+            dest="ignore",
+            required=False,
+            nargs="*",
+            default=[],
+            help="List of taxon ids to be ignored (along with their descendants)",
+        )
     parser.add_argument(
         "--rank",
         dest="rank",
@@ -165,15 +187,21 @@ def main():
     if args.rank:
         args.rank = rank_dict[args.rank]
 
-    if not args.splits and not args.rank:
+    if args.ignore and len(args.ignore) > 0:
+        sys.stdout.write("Filtering taxids: [%s]\n" %",".join(args.ignore))
+
+    if not args.splits and not args.rank and not args.ignore:
         args.splits = ["Bacteria", "Viruses", "Metazoa"]
+
+    if args.splits and len(args.splits) > 0:
+        sys.stdout.write("Splitting on: [%s]\n" %",".join(args.splits))
 
     # Start Program
     now = datetime.now()
     time = now.strftime("%m/%d/%Y, %H:%M:%S")
     sys.stdout.write("PROGRAM START TIME: " + time + "\n")
 
-    parse_report_file(args.report_file, args.splits, args.rank, args.save_json)
+    parse_report_file(args.report_file, args.splits, args.rank, args.ignore, args.save_json)
 
     now = datetime.now()
     time = now.strftime("%m/%d/%Y, %H:%M:%S")
