+++ conf/params.config
@@ -94,6 +94,15 @@ params {
     additional_bracken_jsons = null
     default_taxonomy = 'https://ftp.ncbi.nlm.nih.gov/pub/taxonomy/taxdump.tar.gz'
 
+    spike_ins = null
+    spike_in_sets = [
+        'tobacco_mosaic_virus':[12242],
+        'zymo_D6300': [287,562,28901,1613,1351,1280,1639,1423,4932,5207],
+        'zymo_D6331': [853,423477,301301,817,28128,1680,851,1613,1496,239935,2173,28901,1351,1502,562,5476,4932],
+        'zymo_D6320': [1165090,570278],
+        'zymo_D6321': [1165090,570278,332249]
+    ]
+
     kreport_splits = "Bacteria Viruses Metazoa"
     extract_rank="S"
     extract_thresholds = [
